Create Check whether BST contains Dead End->1st Dec  2023
Given a Binary Search Tree that contains unique positive integer values greater than 0. The task is to complete the function isDeadEnd which returns true if the BST contains a dead end else returns false. Here Dead End means a leaf node, at which no other node can be inserted. that is to check if the leaf node x, x-1 and x+1 is present or not